+++ R/create_waiting_list.R
@@ -42,7 +42,7 @@
 create_waiting_list <- function(n, mean_arrival_rate, mean_wait,
                                 start_date = Sys.Date(), limit_removals = TRUE,
                                 sd = 0, rott = 0, ...) {
-  if (! n > 0) {
+  if (!n > 0) {
     stop("Invalid rate (n) value supplied")
   }
 
@@ -70,7 +70,7 @@ create_waiting_list <- function(n, mean_arrival_rate, mean_wait,
   # simulate non-zero waitlist at end of simulated period.
   if (limit_removals) {
     test_df$removal_date[test_df$removal_date >
-                           (as.Date(start_date, format = "%Y-%m-%d") + n)] <- NA
+      (as.Date(start_date, format = "%Y-%m-%d") + n)] <- NA
     test_df$wait_length[is.na(test_df$removal_date)] <- NA
   }