+++ R/wl_stats.R
@@ -116,8 +116,10 @@ wl_stats <- function(waiting_list,
 
   # load
   q_load <-
-    calc_queue_load(referral_stats$demand_weekly
-                    , removal_stats$capacity_weekly)
+    calc_queue_load(
+      referral_stats$demand_weekly,
+      removal_stats$capacity_weekly
+    )
 
   # load too big
   q_load_too_big <- (q_load >= 1.)
@@ -179,5 +181,4 @@ wl_stats <- function(waiting_list,
   )
 
   return(waiting_stats)
-
 }