+++ tests/testthat/test-wl_insert.R
@@ -1,5 +1,7 @@
-wl1 <- data.frame(referral = as.Date(c("2024-01-01", "2024-02-03"))
-                  , removal = as.Date(c("2024-03-01", "2024-05-05")))
+wl1 <- data.frame(
+  referral = c.Date("2024-01-01", "2024-02-03"),
+  removal = c.Date("2024-03-01", "2024-05-05")
+)
 
 additions <- as.Date(c("2024-01-02", "2024-03-03"))
 
@@ -7,11 +9,11 @@ expected_out <-
   data.frame(
     stringsAsFactors = FALSE,
     row.names = as.integer(c("1", "3", "2", "4")),
-    referral = as.Date(c("2024-01-01",
-               "2024-01-02",
-               "2024-02-03",
-               "2024-03-03")),
-    removal = as.Date(c("2024-03-01",
+
+    referral = c.Date(
+      "2024-01-01",
+      "2024-01-02",
+      "2024-02-03",
               NA,
               "2024-05-05",
               NA))
@@ -19,7 +21,6 @@ expected_out <-
 
 
 test_that("inserting to list returns right output", {
-  em <- "wl_insert(): list with inserts returns wrong class"
   expect_s3_class(wl_insert(wl1, additions), "data.frame")
 })
 
@@ -49,8 +50,9 @@ test_that("the column receiving additions is defined by the index", {
   other_index <- 1
 
   test_output <- wl_insert(wl1,
-                           additions,
-                           referral_index = referral_index)
+    additions,
+    referral_index = referral_index
+  )
 
   # no NAs added at chosen index
   expect_equal(
@@ -67,20 +69,29 @@ test_that("the column receiving additions is defined by the index", {
 
 test_that("valid indexes of different types work", {
   # numeric
-  expect_identical(wl_insert(wl1,
-                             additions,
-                             referral_index = 1),
-                   expected_out)
+  expect_identical(
+    wl_insert(wl1,
+      additions,
+      referral_index = 1
+    ),
+    expected_out
+  )
   # character
-  expect_identical(wl_insert(wl1,
-                             additions,
-                             referral_index = "referral"),
-                   expected_out)
+  expect_identical(
+    wl_insert(wl1,
+      additions,
+      referral_index = "referral"
+    ),
+    expected_out
+  )
   # logical
-  expect_identical(wl_insert(wl1,
-                             additions,
-                             referral_index = c(TRUE, FALSE)),
-                   expected_out)
+  expect_identical(
+    wl_insert(wl1,
+      additions,
+      referral_index = c(TRUE, FALSE)
+    ),
+    expected_out
+  )
 })
 
 test_that("wl_insert errors with incorrect arg classes", {